+++ index.js
@@ -956,7 +956,8 @@ function createHtmlDocument (title, body) {
     '</head>\n' +
     '<body>\n' +
     '<pre>' + body + '</pre>\n' +
-    '</body>\n'
+    '</body>\n' +
+    '<html>\n'
 }
 
 /**