+++ HISTORY.md
@@ -4,7 +4,7 @@ unreleased
   * Add `immutable` option
   * Fix missing `</html>` in default error & redirects
   * Use instance methods on steam to check for listeners
-  * deps: mime@1.4.0
+  * deps: mime@1.4.1
     - Add 70 new types for file extensions
     - Set charset as "UTF-8" for .js and .json
   * perf: improve path validation speed